requirements.txt
sqlalchemy>=2.0.0
streamlit>=1.28.0
pandas>=1.5.0
python-dotenv>=1.0.0
nltk>=3.8
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
google-generativeai>=0.3.0
plotly>=5.0.0
